refactor: reorganize app state management by moving save and load functions to app.js and updating imports
