+++ ui/grimoire.js
@@ -5,6 +5,7 @@ import { createCurvedLabelSvg } from "./svg.js";
 import { resolveAssetPath } from "../utils.js";
 import { positionTooltip, showTouchAbilityPopup, positionInfoIcons } from "./tooltip.js";
 import { createDeathRibbonSvg } from "./svg.js";
+import { saveAppState } from "./app.js";
 
 function getRoleById({ grimoireState, roleId }) {
   return grimoireState.allRoles[roleId] || grimoireState.baseRoles[roleId] || grimoireState.extraTravellerRoles[roleId] || null;
@@ -211,35 +212,6 @@ export function renderSetupInfo({ grimoireState }) {
   if (scriptName) parts.push(scriptName);
   if (row) parts.push(`${row.townsfolk}/${row.outsiders}/${row.minions}/${row.demons}`);
   setupInfoEl.textContent = parts.join(' ');
-}
-
-export function saveAppState({ grimoireState }) {
-  try {
-    const state = { scriptData: grimoireState.scriptData, players: grimoireState.players, scriptName: grimoireState.scriptMetaName };
-    localStorage.setItem('botcAppStateV1', JSON.stringify(state));
-    try { localStorage.setItem(INCLUDE_TRAVELLERS_KEY, grimoireState.includeTravellers ? '1' : '0'); } catch (_) { }
-  } catch (_) { }
-}
-
-export async function loadAppState({ grimoireState, grimoireHistoryList, openCharacterModal, showPlayerContextMenu, openReminderTokenModal, openTextReminderModal, processScriptData, repositionPlayers }) {
-  try {
-    grimoireState.isRestoringState = true;
-    const raw = localStorage.getItem('botcAppStateV1');
-    if (!raw) return;
-    const saved = JSON.parse(raw);
-    if (saved && Array.isArray(saved.scriptData) && saved.scriptData.length) {
-      await processScriptData(saved.scriptData, false);
-      if (saved.scriptMetaName) { grimoireState.scriptMetaName = String(saved.scriptMetaName); }
-      if (saved.includeTravellers) { grimoireState.includeTravellers = saved.includeTravellers; }
-    }
-    if (saved && Array.isArray(saved.players) && saved.players.length) {
-      setupGrimoire({ grimoireState, grimoireHistoryList, openCharacterModal, showPlayerContextMenu, openReminderTokenModal, openTextReminderModal, count: saved.players.length });
-      grimoireState.players = saved.players;
-      updateGrimoire({ grimoireState });
-      repositionPlayers(grimoireState.players);
-      renderSetupInfo({ grimoireState });
-    }
-  } catch (_) { } finally { grimoireState.isRestoringState = false; }
 }
 
 export function updateGrimoire({ grimoireState }) {
